Add search to the Landing Page
Show the search icon centred beneath the Mode Cards, with the same open,
close and keyboard behaviour as the Visualiser. The Landing Page has no
current chat to scope to, so it hides the mode tabs and always searches
all chats, and picking a result enters the Visualiser in the mode that
owns the matching file.

Two changes were needed to make that work:

- performSearch bailed out whenever allMessages was empty, which is
  always the case on the Landing Page. That guard now applies only to
  the current-chat path; the all-chats path reads chatFiles directly.
- scrollToMessage was a silent no-op when the tree had not rendered the
  node yet. It now reports whether it found the node so the Landing Page
  can retry until the tree mounts after the View switch.

Move the fileserver sync from Sidebar into ChatContext so it also runs on
the Landing Page, where search needs the latest files. The Sidebar's sync
button and spinner now share that single implementation. The automatic
sync runs silently so a bad password or a failed request cannot throw
alerts at someone sitting on the Landing Page.

Give the Mode Cards a fixed matching width so they stay symmetric about
the same axis as the title and the search icon.
